store와 slice형태로 파일 구분 1. store.js 파일 추가 2. counterslice.js 파일 추가
